soc: qcom: service-locator: Fix kernel panic in pd_locator_work()
The memory allocated to data is not zeroed out, so it can contain junk. In
the case when service_locator_send_msg() doesn't get a chance to set
domain_list to NULL, domain_list can contain junk which would then be
passed to kfree(), thus causing a kernel panic.

Fix the kernel panic by ensuring that domain_list is always initialized to
NULL before it can be passed to kfree().

Change-Id: I89103d83573c57e819b3f7824c025aa68eb345d8
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>