msm: mdss: Speed up mdss_mdp_get_format_params for UBWC formats
In mdss_mdp_get_format_params, a very large array (mdss_mdp_format_map)
is scanned for a requested format first, and then when the format is
not found in the large array, a small array (mdss_mdp_format_ubwc_map)
is scanned for the requested format. This results in significant
overhead from needlessly scanning the large array for UBWC formats.

To fix this, scan the much-smaller mdss_mdp_format_ubwc_map array first
to search for a requested format.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>