carbon_wahoo_defconfig: Bump to 1000 hertz
In my own testing, battery life was barely impacted (if at all), while the
overall responsiveness of the device was significantly increased. It is my
opinion that such a tradeoff is worth is worth it in the end.

I understand it was decreased to increase throughput and battery life. While
both cases are certainly great reasons to leave it at 100 Hz, I didn't
observe increased battery life to hold true granted one's use case may differ
a lot from my own.

Upping the tick rate might appear counterproductive, but as per Google:

"Historically, the value has been set to 100 on arm and arm64 platforms.
However, this is an accident of history and is not a good value to use for
interactive devices. CONFIG_HZ=100 means that a jiffy is 10ms long, which
means that load balancing between CPUs may take 20ms (two jiffies) to
happen. This can significantly contribute to jank on a loaded system."

But why 1000 when Google themselves uses 300? The issue with using a value
such as 300 is the possibility of things such as poor task placement due to
rounding errors... And well, Google likes to contradict themselves:

"Finally, we've experimented with CONFIG_HZ=1000 on Nexus/Pixel and found it
offers a noticeable performance and power reduction due to decreased RCU overhead"

This material was pulled from here:

https://source.android.com/devices/tech/debug/jank_jitter

Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>