fs: sync: Implement suspend based fsync
Android overuses fsync for several operations however, it is not
desirable to disable it completely.

This implementation uses a bool labelled fsync_on_suspend to determine
whether or not to consider the display state when adjusting fsync. If
disabled, the state of fsync_enabled will be given priority.

If fsync_enabled is disabled, the state of fsync_on_suspend will not
be considered, and fsync will be disabled at all times.

Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>