input: uinput: Enable events from hall IC
This sensor is required for lift to wake, and without it, double tap
to wake missfires in my pocket when I'm out walking.

I can see how the sensor could be pretty annoying (it is), but disabling
it breaks functionality that users expect.

This can still be disabled if desired.

Signed-off-by: RJ Trujillo <certifiedblyndguy@gmail.com>